Former-commit-id: b6f272e37e1675abc18f41641637f67028e7c0dd
Former-commit-id: 54a40439addaac59b3111ef4dcd55e9bfa00eff3

+++ pre-registration/pre-registration-document-service/src/main/java/io/mosip/preregistration/documents/service/DocumentService.java
@@ -194,7 +194,8 @@ public class DocumentService {
 		MainResponseDTO<DocumentResponseDTO> responseDto = new MainResponseDTO<>();
 		MainRequestDTO<DocumentRequestDTO> docReqDto = null;
 		boolean isUploadSuccess = false;
-
+		responseDto.setId(uploadId);
+		responseDto.setVersion(ver);
 		try {
 			docReqDto = serviceUtil.createUploadDto(documentJsonString, preRegistrationId);
 			responseDto.setId(docReqDto.getId());
@@ -407,11 +408,11 @@ public class DocumentService {
 		responseDto.setId(fetchMetaDataId);
 		responseDto.setVersion(ver);
 		boolean isRetrieveSuccess = false;
-		boolean isDocNotFound=false;
+		boolean isDocNotFound = false;
 		Map<String, String> requestParamMap = new HashMap<>();
 		try {
 			requestParamMap.put(RequestCodes.PRE_REGISTRATION_ID, preId);
-			if (ValidationUtil.requstParamValidator(requestParamMap)) {
+			if (ValidationUtil.requstParamValidator(requestParamMap) && serviceUtil.getPreRegInfoRestService(preId)) {
 				List<DocumentEntity> documentEntities = documnetDAO.findBypreregId(preId);
 				responseDto.setResponse(createDocumentResponse(documentEntities));
 				responseDto.setResponsetime(serviceUtil.getCurrentResponseTime());
@@ -421,8 +422,8 @@ public class DocumentService {
 		} catch (Exception ex) {
 			log.error("sessionId", "idType", "id",
 					"In getAllDocumentForPreId method of document service - " + ex.getMessage());
-			if(ex instanceof DocumentNotFoundException)
-				isDocNotFound=true;
+			if (ex instanceof DocumentNotFoundException)
+				isDocNotFound = true;
 			new DocumentExceptionCatcher().handle(ex, responseDto);
 		} finally {
 			if (isRetrieveSuccess) {
@@ -430,15 +431,16 @@ public class DocumentService {
 						"Retrieval of document is successfull", AuditLogVariables.MULTIPLE_ID.toString(),
 						authUserDetails().getUserId(), authUserDetails().getUsername());
 			} else {
-				if(isDocNotFound) {
-					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-							"No documents found for the application", AuditLogVariables.NO_ID.toString(),
-							authUserDetails().getUserId(), authUserDetails().getUsername());
-				}
-				else {
-				setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-						"Retrieval of document is failed", AuditLogVariables.NO_ID.toString(),
-						authUserDetails().getUserId(), authUserDetails().getUsername());
+				if (isDocNotFound) {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "No documents found for the application",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
+				} else {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "Retrieval of document is failed",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
 				}
 			}
 		}
@@ -461,11 +463,11 @@ public class DocumentService {
 		responseDto.setId(fetchContentId);
 		responseDto.setVersion(ver);
 		boolean isRetrieveSuccess = false;
-		boolean isDocNotFound=false;
+		boolean isDocNotFound = false;
 		Map<String, String> requestParamMap = new HashMap<>();
 		try {
 			requestParamMap.put(RequestCodes.PRE_REGISTRATION_ID, preId);
-			if (ValidationUtil.requstParamValidator(requestParamMap)) {
+			if (ValidationUtil.requstParamValidator(requestParamMap) && serviceUtil.getPreRegInfoRestService(preId)) {
 				DocumentEntity documentEntity = documnetDAO.findBydocumentId(docId);
 				if (!documentEntity.getPreregId().equals(preId)) {
 					throw new InvalidDocumentIdExcepion(ErrorCodes.PRG_PAM_DOC_022.name(),
@@ -495,8 +497,8 @@ public class DocumentService {
 		} catch (Exception ex) {
 			log.error("sessionId", "idType", "id",
 					"In getAllDocumentForPreId method of document service - " + ex.getMessage());
-			if(ex instanceof DocumentNotFoundException)
-				isDocNotFound=true;
+			if (ex instanceof DocumentNotFoundException)
+				isDocNotFound = true;
 			new DocumentExceptionCatcher().handle(ex, responseDto);
 		} finally {
 			if (isRetrieveSuccess) {
@@ -504,15 +506,16 @@ public class DocumentService {
 						"Retrieval of document is successfull", AuditLogVariables.MULTIPLE_ID.toString(),
 						authUserDetails().getUserId(), authUserDetails().getUsername());
 			} else {
-				if(isDocNotFound) {
-					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-							"No documents found for the application", AuditLogVariables.NO_ID.toString(),
-							authUserDetails().getUserId(), authUserDetails().getUsername());
-				}
-				else {
-				setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-						"Retrieval of document is failed", AuditLogVariables.NO_ID.toString(),
-						authUserDetails().getUserId(), authUserDetails().getUsername());
+				if (isDocNotFound) {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "No documents found for the application",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
+				} else {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "Retrieval of document is failed",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
 				}
 			}
 		}
@@ -557,52 +560,56 @@ public class DocumentService {
 		delResponseDto.setId(deleteSpecificId);
 		delResponseDto.setVersion(ver);
 		boolean isRetrieveSuccess = false;
-		boolean isDocNotFound=false;
+		boolean isDocNotFound = false;
+		Map<String, String> requestParamMap = new HashMap<>();
 		try {
-			DocumentEntity documentEntity = documnetDAO.findBydocumentId(documentId);
-			if (!documentEntity.getPreregId().equals(preRegistrationId)) {
-				throw new InvalidDocumentIdExcepion(ErrorCodes.PRG_PAM_DOC_022.name(),
-						ErrorMessages.INVALID_DOCUMENT_ID.getMessage());
-			}
-			if (documnetDAO.deleteAllBydocumentId(documentId) > 0) {
-				String key = documentEntity.getDocCatCode() + "_" + documentEntity.getDocumentId();
-				boolean isDeleted = fs.deleteFile(documentEntity.getPreregId(), key);
-				if (!isDeleted) {
-					throw new FSServerException(ErrorCodes.PRG_PAM_DOC_006.toString(),
-							ErrorMessages.DOCUMENT_FAILED_TO_DELETE.getMessage());
+			requestParamMap.put(RequestCodes.PRE_REGISTRATION_ID, preRegistrationId);
+			if (ValidationUtil.requstParamValidator(requestParamMap)
+					&& serviceUtil.getPreRegInfoRestService(preRegistrationId)) {
+				DocumentEntity documentEntity = documnetDAO.findBydocumentId(documentId);
+				if (!documentEntity.getPreregId().equals(preRegistrationId)) {
+					throw new InvalidDocumentIdExcepion(ErrorCodes.PRG_PAM_DOC_022.name(),
+							ErrorMessages.INVALID_DOCUMENT_ID.getMessage());
 				}
-				DocumentDeleteResponseDTO deleteDTO = new DocumentDeleteResponseDTO();
-				deleteDTO.setMessage(DocumentStatusMessages.DOCUMENT_DELETE_SUCCESSFUL.getMessage());
-				delResponseDto.setResponse(deleteDTO);
+				if (documnetDAO.deleteAllBydocumentId(documentId) > 0) {
+					String key = documentEntity.getDocCatCode() + "_" + documentEntity.getDocumentId();
+					boolean isDeleted = fs.deleteFile(documentEntity.getPreregId(), key);
+					if (!isDeleted) {
+						throw new FSServerException(ErrorCodes.PRG_PAM_DOC_006.toString(),
+								ErrorMessages.DOCUMENT_FAILED_TO_DELETE.getMessage());
+					}
+					DocumentDeleteResponseDTO deleteDTO = new DocumentDeleteResponseDTO();
+					deleteDTO.setMessage(DocumentStatusMessages.DOCUMENT_DELETE_SUCCESSFUL.getMessage());
+					delResponseDto.setResponse(deleteDTO);
+				}
 			}
-			
 			delResponseDto.setResponsetime(serviceUtil.getCurrentResponseTime());
-			isRetrieveSuccess=true;
+			isRetrieveSuccess = true;
 
 		} catch (Exception ex) {
 			log.error("sessionId", "idType", "id", "In deleteDocument method of document service - " + ex.getMessage());
-			if(ex instanceof DocumentNotFoundException)
-				isDocNotFound=true;
+			if (ex instanceof DocumentNotFoundException)
+				isDocNotFound = true;
 			new DocumentExceptionCatcher().handle(ex, delResponseDto);
-		}
-		 finally {
-				if (isRetrieveSuccess) {
-					setAuditValues(EventId.PRE_403.toString(), EventName.DELETE.toString(), EventType.BUSINESS.toString(),
-							"Document successfully deleted from the document table", AuditLogVariables.MULTIPLE_ID.toString(),
-							authUserDetails().getUserId(), authUserDetails().getUsername());
+		} finally {
+			if (isRetrieveSuccess) {
+				setAuditValues(EventId.PRE_403.toString(), EventName.DELETE.toString(), EventType.BUSINESS.toString(),
+						"Document successfully deleted from the document table",
+						AuditLogVariables.MULTIPLE_ID.toString(), authUserDetails().getUserId(),
+						authUserDetails().getUsername());
+			} else {
+				if (isDocNotFound) {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "No documents found for the application",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
 				} else {
-					if(isDocNotFound) {
-						setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-								"No documents found for the application", AuditLogVariables.NO_ID.toString(),
-								authUserDetails().getUserId(), authUserDetails().getUsername());
-					}
-					else {
-					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-							"Document deletion failed", AuditLogVariables.NO_ID.toString(),
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "Document deletion failed", AuditLogVariables.NO_ID.toString(),
 							authUserDetails().getUserId(), authUserDetails().getUsername());
-					}
 				}
-		 }
+			}
+		}
 		return delResponseDto;
 	}
 
@@ -617,14 +624,15 @@ public class DocumentService {
 	public MainResponseDTO<DocumentDeleteResponseDTO> deleteAllByPreId(String preregId) {
 		log.info("sessionId", "idType", "id", "In deleteAllByPreId method of document service");
 		boolean isDeleteSuccess = false;
-		boolean isDocNotFound=false;
+		boolean isDocNotFound = false;
 		MainResponseDTO<DocumentDeleteResponseDTO> deleteRes = new MainResponseDTO<>();
 		deleteRes.setId(deleteId);
 		deleteRes.setVersion(ver);
 		Map<String, String> requestParamMap = new HashMap<>();
 		try {
 			requestParamMap.put(RequestCodes.PRE_REGISTRATION_ID, preregId);
-			if (ValidationUtil.requstParamValidator(requestParamMap)) {
+			if (ValidationUtil.requstParamValidator(requestParamMap)
+					&& serviceUtil.getPreRegInfoRestService(preregId)) {
 				List<DocumentEntity> documentEntityList = documnetDAO.findBypreregId(preregId);
 				DocumentDeleteResponseDTO deleteDTO = deleteFile(documentEntityList, preregId);
 				deleteRes.setResponse(deleteDTO);
@@ -635,8 +643,8 @@ public class DocumentService {
 		} catch (Exception ex) {
 			log.error("sessionId", "idType", "id",
 					"In deleteAllByPreId method of document service - " + ex.getMessage());
-			if(ex instanceof DocumentNotFoundException)
-				isDocNotFound=true;
+			if (ex instanceof DocumentNotFoundException)
+				isDocNotFound = true;
 			new DocumentExceptionCatcher().handle(ex, deleteRes);
 		} finally {
 			if (isDeleteSuccess) {
@@ -645,15 +653,15 @@ public class DocumentService {
 						AuditLogVariables.MULTIPLE_ID.toString(), authUserDetails().getUserId(),
 						authUserDetails().getUsername());
 			} else {
-				if(isDocNotFound) {
-					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-							"No documents found for the application", AuditLogVariables.NO_ID.toString(),
+				if (isDocNotFound) {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "No documents found for the application",
+							AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
+							authUserDetails().getUsername());
+				} else {
+					setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(),
+							EventType.SYSTEM.toString(), "Document deletion failed", AuditLogVariables.NO_ID.toString(),
 							authUserDetails().getUserId(), authUserDetails().getUsername());
-				}
-				else {
-				setAuditValues(EventId.PRE_405.toString(), EventName.EXCEPTION.toString(), EventType.SYSTEM.toString(),
-						"Document deletion failed", AuditLogVariables.NO_ID.toString(), authUserDetails().getUserId(),
-						authUserDetails().getUsername());
 				}
 			}
 		}
